+++ src/main/java/org/folio/inventory/dataimport/exceptions/DataImportException.java
@@ -1,0 +1,8 @@
+package org.folio.inventory.dataimport.exceptions;
+
+public class DataImportException extends RuntimeException {
+
+  public DataImportException(String message) {
+    super(message);
+  }
+}